refactor out p2p client for op-services/sources+apis
